#281 [chore] persistence 패키지 구성
DB와 연관된 영속성 계층과 관련된 파일 위주로 구성했습니다
