--- src/main/java/com/asap/server/domain/PreferTime.java
@@ -1,42 +1,0 @@
-package com.asap.server.domain;
-
-import com.asap.server.domain.enums.TimeSlot;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PreferTime extends AuditingTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "meeting_id")
-    private Meeting meeting;
-
-    @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private TimeSlot startTime;
-
-    @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private TimeSlot endTime;
-
-}
