--- src/main/java/com/asap/server/domain/User.java
@@ -1,46 +1,0 @@
-package com.asap.server.domain;
-
-import com.asap.server.domain.enums.Role;
-import jakarta.persistence.FetchType;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.ColumnDefault;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class User extends AuditingTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "meeting_id")
-    private Meeting meeting;
-
-    @Column(nullable = false)
-    private String name;
-
-    @ColumnDefault(value = "false")
-    private Boolean isFixed;
-
-    @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private Role role;
-}
