--- src/main/java/com/asap/server/domain/enums/Role.java
@@ -1,15 +1,0 @@
-package com.asap.server.domain.enums;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-public enum Role {
-    HOST("HOST"),
-    MEMBER("MEMBER");
-
-    @Getter
-    @JsonValue
-    private String role;
-}
