+++ src/main/java/com/asap/server/persistence/repository/TimeBlockUserRepository.java
@@ -1,0 +1,17 @@
+package com.asap.server.persistence.repository;
+
+import com.asap.server.persistence.domain.TimeBlock;
+import com.asap.server.persistence.domain.TimeBlockUser;
+import com.asap.server.persistence.domain.User;
+import org.springframework.data.repository.Repository;
+
+import java.util.List;
+
+public interface TimeBlockUserRepository extends Repository<TimeBlockUser, Long> {
+
+    void save(final TimeBlockUser timeBlockUser);
+
+    List<TimeBlockUser> findAllByUser(final User user);
+
+    List<TimeBlockUser> findByTimeBlock(final TimeBlock timeBlock);
+}
