+++ src/main/java/com/asap/server/persistence/repository/timeblock/TimeBlockRepositoryCustom.java
@@ -1,0 +1,8 @@
+package com.asap.server.persistence.repository.timeblock;
+
+import com.asap.server.persistence.repository.timeblock.dto.TimeBlockDto;
+import java.util.List;
+
+public interface TimeBlockRepositoryCustom {
+    List<TimeBlockDto> findAllTimeBlockByMeeting(final Long meetingId);
+}
