--- src/main/java/com/asap/server/repository/meeting/MeetingRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.asap.server.repository.meeting;
-
-import com.asap.server.domain.Meeting;
-
-import java.util.Optional;
-
-public interface MeetingRepositoryCustom {
-    Optional<Meeting> findByIdWithHost(final Long id);
-}
