--- src/main/java/com/asap/server/repository/timeblock/TimeBlockRepository.java
@@ -1,18 +1,0 @@
-package com.asap.server.repository.timeblock;
-
-import com.asap.server.domain.AvailableDate;
-import com.asap.server.domain.TimeBlock;
-import com.asap.server.domain.enums.TimeSlot;
-import org.springframework.data.repository.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface TimeBlockRepository extends Repository<TimeBlock, Long>, TimeBlockRepositoryCustom {
-
-    void save(final TimeBlock timeBlock);
-
-    List<TimeBlock> findByAvailableDate(final AvailableDate availableDate);
-
-    Optional<TimeBlock> findByAvailableDateAndTimeSlot(final AvailableDate availableDate, TimeSlot timeSlot);
-}
