--- src/main/java/com/asap/server/repository/timeblock/TimeBlockRepositoryCustom.java
@@ -1,8 +1,0 @@
-package com.asap.server.repository.timeblock;
-
-import com.asap.server.repository.timeblock.dto.TimeBlockDto;
-import java.util.List;
-
-public interface TimeBlockRepositoryCustom {
-    List<TimeBlockDto> findAllTimeBlockByMeeting(final Long meetingId);
-}
