--- src/main/java/com/asap/server/repository/user/UserRepositoryCustom.java
@@ -1,14 +1,0 @@
-package com.asap.server.repository.user;
-
-import com.asap.server.domain.Meeting;
-import com.asap.server.domain.enums.TimeSlot;
-import com.asap.server.service.vo.UserVo;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public interface UserRepositoryCustom {
-    void updateUserIsFixedByMeeting(final Meeting meeting, final List<Long> users);
-
-    List<UserVo> findByAvailableDateAndTimeSlots(Long meetingId, LocalDate availableDate, List<TimeSlot> timeSlots);
-}
