--- src/test/java/com/asap/server/repository/MeetingRepositoryCustomTest.java
@@ -1,68 +1,0 @@
-package com.asap.server.repository;
-
-import com.asap.server.config.querydsl.QueryDslConfig;
-import com.asap.server.domain.Meeting;
-import com.asap.server.domain.Place;
-import com.asap.server.domain.User;
-import com.asap.server.domain.enums.Duration;
-import com.asap.server.domain.enums.PlaceType;
-import com.asap.server.domain.enums.Role;
-import com.asap.server.repository.meeting.MeetingRepository;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
-
-import jakarta.persistence.EntityManager;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-@DataJpaTest
-@Import(QueryDslConfig.class)
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class MeetingRepositoryCustomTest {
-    @Autowired
-    private MeetingRepository meetingRepository;
-
-    @Autowired
-    private EntityManager em;
-
-    @Test
-    @DisplayName("방장이 로그인을 할 때, 방장 정보도 함께 불러온다.")
-    void fetchJoinTest() {
-        final Place place = Place.builder()
-                .placeType(PlaceType.OFFLINE)
-                .build();
-
-        final Meeting meeting = Meeting.builder()
-                .title("회의 테스트")
-                .password("0000")
-                .additionalInfo("")
-                .duration(Duration.HALF)
-                .place(place)
-                .build();
-
-        final User user = User.builder()
-                .meeting(meeting)
-                .name("강원용")
-                .role(Role.HOST)
-                .isFixed(false)
-                .build();
-        meeting.setHost(user);
-
-        em.persist(meeting);
-        em.persist(user);
-        em.flush();
-        em.clear();
-
-        // when
-        Meeting result = meetingRepository.findByIdWithHost(meeting.getId()).get();
-        User host = result.getHost();
-
-        // then
-        assertThat(host).isNotNull();
-    }
-
-}
